Implement native support StringView for contains function
